// packages/core/src/enums/state-name.ts
export enum StateName {
  INIT = "INIT",
  OPEN_JOB_PAGE = "OPEN_JOB_PAGE",
  DETECT_APPLY_ENTRY = "DETECT_APPLY_ENTRY",
  LOGIN_OR_CONTINUE = "LOGIN_OR_CONTINUE",
  UPLOAD_RESUME = "UPLOAD_RESUME",
  WAIT_FOR_PARSE = "WAIT_FOR_PARSE",
  VALIDATE_PARSED_PROFILE = "VALIDATE_PARSED_PROFILE",
  FILL_REQUIRED_FIELDS = "FILL_REQUIRED_FIELDS",
  ANSWER_SCREENING_QUESTIONS = "ANSWER_SCREENING_QUESTIONS",
  REVIEW_DISCLOSURES = "REVIEW_DISCLOSURES",
  PRE_SUBMIT_CHECK = "PRE_SUBMIT_CHECK",
  SUBMIT = "SUBMIT",
  CAPTURE_CONFIRMATION = "CAPTURE_CONFIRMATION",
  ESCALATE = "ESCALATE",
}